Add custom python.el
Commit 1808d254a5820e8c650013033f800ca2990cd239 to the official emacs
mirror on GitHub (https://github.com/emacs-mirror/emacs) broke python
mode for emacs 24.5.1. Since Emacs 25.1 doesn't seem to be available
in my apt repositories, I added a custom python.el that reversed the
changes in this commit. It functions as necessary and we get the
added functionality of Python3 keywords.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch master
# Your branch is up-to-date with 'origin/master'.
#
# Changes to be committed:
#	new file:   .gitignore
#	new file:   python.el
#